+++ pkg/store/postings_codec_test.go
@@ -4,13 +4,16 @@
 package store
 
 import (
+	"bytes"
 	"context"
+	crand "crypto/rand"
 	"math"
 	"math/rand"
 	"sort"
 	"strconv"
 	"testing"
 
+	"github.com/klauspost/compress/s2"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
@@ -19,6 +22,50 @@ import (
 	"github.com/efficientgo/core/testutil"
 	storetestutil "github.com/thanos-io/thanos/pkg/store/storepb/testutil"
 )
+
+func TestStreamedSnappyMaximumDecodedLen(t *testing.T) {
+	t.Run("compressed", func(t *testing.T) {
+		b := make([]byte, 100)
+		for i := 0; i < 100; i++ {
+			b[i] = 0x42
+		}
+
+		snappyEncoded := &bytes.Buffer{}
+
+		sw := s2.NewWriter(snappyEncoded, s2.WriterSnappyCompat(), s2.WriterBestCompression())
+
+		_, err := sw.Write(b)
+		testutil.Ok(t, err)
+
+		testutil.Ok(t, sw.Close())
+
+		maxLen, err := maximumDecodedLenSnappyStreamed(snappyEncoded.Bytes())
+		testutil.Ok(t, err)
+		t.Log(maxLen)
+		testutil.Assert(t, maxLen == 100)
+	})
+	t.Run("random", func(t *testing.T) {
+		for i := 10000; i < 30000; i++ {
+			b := make([]byte, i)
+			_, err := crand.Read(b)
+			testutil.Ok(t, err)
+
+			snappyEncoded := &bytes.Buffer{}
+
+			sw := s2.NewWriter(snappyEncoded, s2.WriterSnappyCompat())
+
+			_, err = sw.Write(b)
+			testutil.Ok(t, err)
+
+			testutil.Ok(t, sw.Close())
+
+			maxLen, err := maximumDecodedLenSnappyStreamed(snappyEncoded.Bytes())
+			testutil.Ok(t, err)
+			testutil.Assert(t, maxLen > 100)
+			testutil.Assert(t, maxLen < 30000)
+		}
+	})
+}
 
 func TestDiffVarintCodec(t *testing.T) {
 	chunksDir := t.TempDir()
@@ -56,9 +103,11 @@ func TestDiffVarintCodec(t *testing.T) {
 
 	codecs := map[string]struct {
 		codingFunction   func(index.Postings, int) ([]byte, error)
-		decodingFunction func([]byte) (closeablePostings, error)
+		decodingFunction func([]byte, bool) (closeablePostings, error)
 	}{
-		"raw":            {codingFunction: diffVarintEncodeNoHeader, decodingFunction: func(bytes []byte) (closeablePostings, error) { return newDiffVarintPostings(bytes, nil), nil }},
+		"raw": {codingFunction: diffVarintEncodeNoHeader, decodingFunction: func(bytes []byte, disablePooling bool) (closeablePostings, error) {
+			return newDiffVarintPostings(bytes, nil), nil
+		}},
 		"snappy":         {codingFunction: diffVarintSnappyEncode, decodingFunction: diffVarintSnappyDecode},
 		"snappyStreamed": {codingFunction: diffVarintSnappyStreamedEncode, decodingFunction: diffVarintSnappyStreamedDecode},
 	}
@@ -81,7 +130,7 @@ func TestDiffVarintCodec(t *testing.T) {
 				t.Log("encoded size", len(data), "bytes")
 				t.Logf("ratio: %0.3f", float64(len(data))/float64(4*p.len()))
 
-				decodedPostings, err := codec.decodingFunction(data)
+				decodedPostings, err := codec.decodingFunction(data, false)
 				testutil.Ok(t, err)
 
 				p.reset()
@@ -212,21 +261,21 @@ func BenchmarkPostingsEncodingDecoding(b *testing.B) {
 
 	codecs := map[string]struct {
 		codingFunction   func(index.Postings, int) ([]byte, error)
-		decodingFunction func([]byte) (closeablePostings, error)
+		decodingFunction func([]byte, bool) (closeablePostings, error)
 	}{
-		"raw":            {codingFunction: diffVarintEncodeNoHeader, decodingFunction: func(bytes []byte) (closeablePostings, error) { return newDiffVarintPostings(bytes, nil), nil }},
+		"raw": {codingFunction: diffVarintEncodeNoHeader, decodingFunction: func(bytes []byte, disablePooling bool) (closeablePostings, error) {
+			return newDiffVarintPostings(bytes, nil), nil
+		}},
 		"snappy":         {codingFunction: diffVarintSnappyEncode, decodingFunction: diffVarintSnappyDecode},
 		"snappyStreamed": {codingFunction: diffVarintSnappyStreamedEncode, decodingFunction: diffVarintSnappyStreamedDecode},
 	}
-
 	b.ReportAllocs()
 
 	for _, count := range []int{10000, 100000, 1000000} {
-		for codecName, codecFns := range codecs {
-			b.Run(strconv.Itoa(count), func(b *testing.B) {
+		b.Run(strconv.Itoa(count), func(b *testing.B) {
+			for codecName, codecFns := range codecs {
 				b.Run(codecName, func(b *testing.B) {
 					b.Run("encode", func(b *testing.B) {
-						b.ResetTimer()
 						for i := 0; i < b.N; i++ {
 							ps := &uint64Postings{vals: p[:count]}
 
@@ -243,19 +292,23 @@ func BenchmarkPostingsEncodingDecoding(b *testing.B) {
 						if err != nil {
 							b.Fatal(err)
 						}
-
 						b.ResetTimer()
+
 						for i := 0; i < b.N; i++ {
-							_, err := codecFns.decodingFunction(encoded)
+							decoded, err := codecFns.decodingFunction(encoded, true)
 							if err != nil {
 								b.Fatal(err)
 							}
+
+							for decoded.Next() {
+								var _ = decoded.At()
+							}
+							testutil.Ok(b, decoded.Err())
 						}
 					})
-
 				})
-			})
-		}
+			}
+		})
 	}
 }